Revise code and add fixed-point operators to eliminate certain assertion failures.
darcs-hash:20071018182431-d1b56-8b81c16730cb3e5fda4c04daed2d329d711d8b4f.gz

+++ d2/point.h
@@ -133,7 +133,7 @@ public:
 		ale_pos to_p = lengthtosq(p);
 		ale_pos to_q = lengthtosq(q);
 
-		ale_pos cos_of = (to_p + to_q - p.lengthtosq(q))
+		ale_pos cos_of = (double) (to_p + to_q - p.lengthtosq(q))
 			       / (2 * sqrt(to_p) * sqrt(to_q));
 
 		return acos(cos_of);
